Cygwin IP Address Commit.
This commit adds handling to the prompt to handle the fact that when
running under Cygwin you need to obtain ip information from ipconfig
instead of ifconfig.

The output of ipconfig is a mess, especially, if you have vmware or vpns
running.  However, I think this first attempt should suffice.