Bash Cleanup Commit.
This commit changes some comments and adds a couple of new Bash goodies.

First up a rather excellent mkdircd function I found while looking for
something unrelated.  Added to .functions.  Call mkdircd with a
directory structure and it will create the directories and change you to
the top level one.

The other goody has been added to the .exports file and writes a
warning to the stdout if a shell process exists with any value over than
0.