Experimental .promt IP Address Commit.
This commit attempts to rectify the various issues I have found when
trying to get the device and IP address from various distros.

I have altered the method used to get the device and IP address for *nix
based systems to parse ip route.

This commit is made so I can test the changes on various vm's.

Also fixed logic issue with regard to eth device test being made last
which resulted in most displays showing we are on a a wired connections
when we are not.