Rework install_dotfiles Function For Directorys
This commit makes a number of initial changes relating to making the
structure of the .dotfiles/ directory mirror the users home directory.

The ideal is to handle directories at various depths.  Now any file or
directory suffixed with .symlink should work.  If directories do not
exist then they will be created with the required parents.

Commiting to provide a base to complete from.