Update Background Configuration
This commit renames the background directory to backgrounds.  The idea
is that I will update the configuration settings to use the file named
wallpaper.jpg as the desktop wallpaper and the lock screen background.

If there is a file in backgrounds called lockscreen.jpg I will use that
instead for the lockscreen.